+++ apps/wallet-mobile/.storybook/storybook.requires.js
@@ -123,8 +123,8 @@ const getStories = () => {
     "./src/features/Discover/useCases/SearchDappInBrowser/SearchDappInBrowserScreen.stories.tsx": require("../src/features/Discover/useCases/SearchDappInBrowser/SearchDappInBrowserScreen.stories.tsx"),
     "./src/features/Discover/useCases/SelectDappFromList/CountDAppsAvailable/CountDAppsAvailable.stories.tsx": require("../src/features/Discover/useCases/SelectDappFromList/CountDAppsAvailable/CountDAppsAvailable.stories.tsx"),
     "./src/features/Discover/useCases/SelectDappFromList/DAppExplorerTabItem/DAppExplorerTabItem.stories.tsx": require("../src/features/Discover/useCases/SelectDappFromList/DAppExplorerTabItem/DAppExplorerTabItem.stories.tsx"),
-    "./src/features/Discover/useCases/SelectDappFromList/DAppItem/DAppItem.stories.tsx": require("../src/features/Discover/useCases/SelectDappFromList/DAppItem/DAppItem.stories.tsx"),
-    "./src/features/Discover/useCases/SelectDappFromList/DAppItem/DAppItemSkeleton.stories.tsx": require("../src/features/Discover/useCases/SelectDappFromList/DAppItem/DAppItemSkeleton.stories.tsx"),
+    "./src/features/Discover/useCases/SelectDappFromList/DAppListItem/DAppItemSkeleton.stories.tsx": require("../src/features/Discover/useCases/SelectDappFromList/DAppListItem/DAppItemSkeleton.stories.tsx"),
+    "./src/features/Discover/useCases/SelectDappFromList/DAppListItem/DAppListItem.stories.tsx": require("../src/features/Discover/useCases/SelectDappFromList/DAppListItem/DAppListItem.stories.tsx"),
     "./src/features/Discover/useCases/SelectDappFromList/DAppTypes/DAppTypes.stories.tsx": require("../src/features/Discover/useCases/SelectDappFromList/DAppTypes/DAppTypes.stories.tsx"),
     "./src/features/Discover/useCases/SelectDappFromList/SelectDappFromListScreen.stories.tsx": require("../src/features/Discover/useCases/SelectDappFromList/SelectDappFromListScreen.stories.tsx"),
     "./src/features/Discover/useCases/SelectDappFromList/WelcomeDApp.stories.tsx": require("../src/features/Discover/useCases/SelectDappFromList/WelcomeDApp.stories.tsx"),
