+++ apps/wallet-mobile/src/features/Discover/common/helpers.ts
@@ -1,3 +1,5 @@
+import Google from '../../../assets/img/dApp/google.png'
+
 export const validUrl = (url: string) => {
   return /^(?:http(s)?:\/\/)?[\w.-]+(?:\.[\w.-]+)+[\w\-._~:/?#[\]@!&',,=.+]+$/g.test(url)
 }
@@ -26,3 +28,27 @@ export const getDomainFromUrl = (url: string) => {
     }
   }
 }
+
+export interface DAppItem {
+  id: string
+  name: string
+  description: string
+  category: string
+  logo: string
+  uri: string
+  origins: string[]
+}
+
+export const GOOGLE_DAPP_ID = 'google_search'
+
+export const getGoogleSearchItem = (searchQuery: string): DAppItem => ({
+  id: GOOGLE_DAPP_ID,
+  name: searchQuery,
+  description: 'Google',
+  category: 'search',
+  logo: Google,
+  uri: `https://www.google.com/search?q=${encodeURIComponent(searchQuery)}`,
+  origins: ['https://www.google.com'],
+})
+
+export const isGoogleSearchItem = (dApp: DAppItem) => dApp.id === GOOGLE_DAPP_ID
